% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/create_database_table.R
\name{create_database_table}
\alias{create_database_table}
\title{create_database_table}
\usage{
create_database_table(workshop_dir)
}
\arguments{
\item{workshop_dir}{Directory of workshop}
}
\value{
Boolean
}
\description{
create_database_table
}

