\name{SDMengine-package}
\alias{SDMengine-package}
\alias{SDMengine}
\docType{package}
\title{
SDMengine
}
\description{
A SDM(Species Distribution Modelling) modelling software
}
\details{
\tabular{ll}{
Package: \tab SDMengine\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-07-03\cr
License: \tab AGPL\cr
}
SDMengine is a part of SDM serial software
}
\author{
Xiaoquan Kong,
Renyan Duan,
Minyi Huang

Maintainer: Xiaoquan Kong <u1mail2me@gmail.com>
}

\keyword{ Species distribution modelling }
